Breakout MoreNav into own component
