+++ src/components/Organisms/Header2024/HeaderNav2024/HeaderNav2024.js
@@ -3,6 +3,7 @@ import PropTypes from 'prop-types';
 
 import Text from '../../../Atoms/Text/Text';
 import BurgerMenu from '../Burger/BurgerMenu';
+import MoreNav from './MoreNav';
 import { breakpointValues } from '../../../../theme/shared/allBreakpoints';
 import { NavHelper, MoreNavPreProcess } from '../../../../utils/navHelper';
 import { InternalLinkHelper } from '../../../../utils/internalLinkHelper';
@@ -19,14 +20,6 @@ import {
   SubNavLink,
   ChevronWrapper,
   NavMetaIcons,
-  // More Nav stuff:
-  MoreNavLink,
-  MoreSubNavMenu,
-  MoreNavItem,
-  MoreNestedSubNavMenu,
-  MoreSubNavItem,
-  MoreNavNestedLink,
-  MoreSubNavLink,
   DonateButtonWrapperBottom
 } from './HeaderNav2024.style';
 
@@ -207,110 +200,18 @@ const HeaderNav = ({
               );
             })}
 
-            {/*
-            ****************************
-            MORE NAV RENDER STARTS HERE:
-            ****************************
-          */}
-
             {/* Only actually render 'More' nav stuff when we've got content */}
             {(!isNotDesktop && processedItems.moreNavGroups.length > 0) && (
-              <MoreNavItem>
-                {/* The 'More' nav button: */}
-                <Text>
-                  <MoreNavLink
-                    href="#"
-                    inline
-                    // As this is purely used to hover-over, and never represents a
-                    // direct link to a page, we can nip any click event in the bud:
-                    onClick={e => { e.preventDefault(); }}
-                    role="button"
-                    aria-haspopup="true"
-                  >
-                    More
-                    <ChevronWrapper>
-                      <img src={menuGroupIcon} alt="Chevron icon" />
-                    </ChevronWrapper>
-                  </MoreNavLink>
-                </Text>
-
-                {/* The Ul to wrap each of the 'More Nav' menu groups */}
-                <MoreSubNavMenu
-                  role="list"
-                  isFocussed={!!isTabFocussed.more}
-                  key="more-nav-ul"
-                >
-
-                  {/* For each item in this menu group:  */}
-                  {processedItems.moreNavGroups.map(child => {
-                    /* Grab the first links properties to use for our parent/button */
-                    const thisFirstChild = child.links[0];
-                    let thisUrl = NavHelper(thisFirstChild);
-                    const relNoopener = (!allowListed(thisUrl) && 'noopener') || null;
-                    const hasSubMenu = child.links && child.links.length > 1;
-                    const hasPopUp = hasSubMenu ? 'true' : null;
-                    thisUrl = InternalLinkHelper(thisUrl);
-
-                    return (
-                      <MoreSubNavItem key={child.title}>
-                        {/* Either the Direct link (for a one-link menu item)
-                      or a 'button' to show the submenu: */}
-                        <MoreNavNestedLink
-                          href={hasPopUp ? '#' : thisUrl}
-                          inline
-                          rel={relNoopener}
-                          aria-haspopup={hasPopUp}
-                          role={hasPopUp ? 'button' : 'link'}
-                          onClick={hasPopUp ? e => toggleSubMenu(e, child.id) : null}
-                          isSubMenuOpen={!!isSubMenuOpen[child.id]}
-                          aria-expanded={!!isSubMenuOpen[child.id]}
-                          onKeyUp={keyPressed(child.id)}
-                        >
-                          {thisFirstChild.title}
-
-                          {hasSubMenu && (
-                            <ChevronWrapper>
-                              <img src={menuGroupIcon} alt="Chevron icon" />
-                            </ChevronWrapper>
-                          )}
-
-                        </MoreNavNestedLink>
-
-                        <>
-                          {hasSubMenu && (
-                            <MoreNestedSubNavMenu
-                              role="list"
-                              isSubMenuOpen={!!isSubMenuOpen[child.id]}
-                            >
-                              {child.links.map(subChild => {
-                                const thisSubUrl = NavHelper(subChild);
-                                return (
-                                  // 'More Nav' sub item:
-                                  <MoreSubNavItem key={thisSubUrl}>
-                                    <MoreSubNavLink
-                                      href={thisSubUrl}
-                                      inline
-                                      role="menuitem"
-                                      // Allows us to avoid using the 'display:none'
-                                      // approach so we can animate properly:
-                                      tabIndex={isSubMenuOpen[child.id] ? '0' : '-1'}
-                                    >
-                                      <Text>
-                                        {subChild.title}
-                                      </Text>
-                                    </MoreSubNavLink>
-                                  </MoreSubNavItem>
-                                );
-                              })}
-                            </MoreNestedSubNavMenu>
-                          )}
-                        </>
-                      </MoreSubNavItem>
-                    );
-                  })}
-                </MoreSubNavMenu>
-
-              </MoreNavItem>
+              <MoreNav
+                isTabFocussed={isTabFocussed}
+                processedItems={processedItems}
+                isSubMenuOpen={isSubMenuOpen}
+                keyPressed={keyPressed}
+                toggleSubMenu={toggleSubMenu}
+                navHelper={NavHelper}
+                allowListed={allowListed}
+                internalLinkHelper={InternalLinkHelper}
+              />
             )}
 
           </NavMenu>
