+++ src/components/Organisms/Header2024/HeaderNav2024/HeaderNav2024.style.js
@@ -1,5 +1,4 @@
-/* eslint-disable no-multiple-empty-lines */
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 
 import Link from '../../../Atoms/Link/Link';
 import hideVisually from '../../../../theme/shared/hideVisually';
@@ -390,106 +389,106 @@ const DonateButtonWrapperBottom = styled.div`
  * **********
  */
 
-const MoreSubNavMenu = styled(SubNavMenu)`
-  @media ${({ theme }) => theme.allBreakpoints('Nav')} {
-    // top: 88px
-  }
-`;
+// const MoreSubNavMenu = styled(SubNavMenu)`
+//   @media ${({ theme }) => theme.allBreakpoints('Nav')} {
+//     // top: 88px
+//   }
+// `;
 
-const MoreNavLink = styled(NavLink)`
-  @media ${({ theme }) => theme.allBreakpoints('Nav')} {
-    :focus,
-    :hover,
-    :focus-within {
-      + ${MoreSubNavMenu} {
-        display: flex;
-      }
-    }
-  }
-`;
+// const MoreNavLink = styled(NavLink)`
+//   @media ${({ theme }) => theme.allBreakpoints('Nav')} {
+//     :focus,
+//     :hover,
+//     :focus-within {
+//       + ${MoreSubNavMenu} {
+//         display: flex;
+//       }
+//     }
+//   }
+// `;
 
-// Use for the 'More' link only
-const MoreNavItem = styled(NavItem)`
-  @media ${({ theme }) => theme.allBreakpoints('Nav')} {
-    :hover,
-    :focus-within,
-    :focus {
-      > ${MoreSubNavMenu} {
-        visibility: visible;
-        opacity: 1;
-        display: flex;
-      }
-    }
+// // Use for the 'More' link only
+// const MoreNavItem = styled(NavItem)`
+//   @media ${({ theme }) => theme.allBreakpoints('Nav')} {
+//     :hover,
+//     :focus-within,
+//     :focus {
+//       > ${MoreSubNavMenu} {
+//         visibility: visible;
+//         opacity: 1;
+//         display: flex;
+//       }
+//     }
 
-    :hover,
-    :focus,
-    :focus-within {
-      ${MoreSubNavMenu} {
-        display: flex;
-        flex-direction: column;
-      }
-    }
-  }
-`;
+//     :hover,
+//     :focus,
+//     :focus-within {
+//       ${MoreSubNavMenu} {
+//         display: flex;
+//         flex-direction: column;
+//       }
+//     }
+//   }
+// `;
 
-// Clone Of SubNavMenu, NESTED menu
-const MoreNestedSubNavMenu = styled(SubNavMenu)`
-  @media ${({ theme }) => theme.allBreakpoints('Nav')} {
-    top: 0;
-    left: -1px;
-    border-radius: 0;
-    border-bottom: 0;
-    position: relative;
-    transition: max-height 0.5s cubic-bezier(0.5, 1.5, 0.5, 0.80);
-    max-height: 0;
-    border: 0;
-    box-shadow: none;
+// // Clone Of SubNavMenu, NESTED menu
+// const MoreNestedSubNavMenu = styled(SubNavMenu)`
+//   @media ${({ theme }) => theme.allBreakpoints('Nav')} {
+//     top: 0;
+//     left: -1px;
+//     border-radius: 0;
+//     border-bottom: 0;
+//     position: relative;
+//     transition: max-height 0.5s cubic-bezier(0.5, 1.5, 0.5, 0.80);
+//     max-height: 0;
+//     border: 0;
+//     box-shadow: none;
 
-    ${({ isSubMenuOpen }) => (isSubMenuOpen && css`
-      // Calculating a realistic height ceiling (8 subnav items) 
-      // to make animation as slick as possible
-      max-height: calc(57px * 8);
-  `)};
-  }
-`;
+//     ${({ isSubMenuOpen }) => (isSubMenuOpen && css`
+//       // Calculating a realistic height ceiling (8 subnav items)
+//       // to make animation as slick as possible
+//       max-height: calc(57px * 8);
+//   `)};
+//   }
+// `;
 
-const MoreSubNavItem = styled(SubNavItem)`
-  // Chevron icon
-  > a > div {
-    position: absolute;
-    width: 15px;
-    top: 13px;
-    right: 25px;
-    transition: transform 0.35s cubic-bezier(0.41, 1.64, 0.41, 0.8);
-  }
+// const MoreSubNavItem = styled(SubNavItem)`
+//   // Chevron icon
+//   > a > div {
+//     position: absolute;
+//     width: 15px;
+//     top: 13px;
+//     right: 25px;
+//     transition: transform 0.35s cubic-bezier(0.41, 1.64, 0.41, 0.8);
+//   }
 
-  :hover, 
-  :focus,
-  :focus-within {
-    > a {
-      color: ${({ theme }) => theme.color('red')};
-    }
-    > a > div {
-      img {
-        filter: invert(0.5) sepia(1) saturate(100) hue-rotate(20deg);
-      }
-    }
-  }
-`;
+//   :hover,
+//   :focus,
+//   :focus-within {
+//     > a {
+//       color: ${({ theme }) => theme.color('red')};
+//     }
+//     > a > div {
+//       img {
+//         filter: invert(0.5) sepia(1) saturate(100) hue-rotate(20deg);
+//       }
+//     }
+//   }
+// `;
 
-const MoreNavNestedLink = styled(NavLink)`
-  padding: 20px 25px 18px;
+// const MoreNavNestedLink = styled(NavLink)`
+//   padding: 20px 25px 18px;
 
-  ${({ isSubMenuOpen }) => (isSubMenuOpen && css`
-    > div {
-      transform: rotate(-180deg);
-    }
-  `)};
-`;
+//   ${({ isSubMenuOpen }) => (isSubMenuOpen && css`
+//     > div {
+//       transform: rotate(-180deg);
+//     }
+//   `)};
+// `;
 
-const MoreSubNavLink = styled(SubNavLink)`
-  //
-`;
+// const MoreSubNavLink = styled(SubNavLink)`
+//   //
+// `;
 
 export {
   Nav,
@@ -501,13 +500,13 @@ export {
   SubNavLink,
   ChevronWrapper,
   NavMetaIcons,
-  DonateButtonWrapperBottom,
+  DonateButtonWrapperBottom
   // 'More Nav' stuff:
-  MoreNavLink,
-  MoreSubNavMenu,
-  MoreNavItem,
-  MoreNestedSubNavMenu,
-  MoreSubNavItem,
-  MoreNavNestedLink,
-  MoreSubNavLink
+  // MoreNavLink,
+  // MoreSubNavMenu,
+  // MoreNavItem,
+  // MoreNestedSubNavMenu,
+  // MoreSubNavItem,
+  // MoreNavNestedLink,
+  // MoreSubNavLink
 };
